requirements.txt
PyGithub
csv
attrs
black
isort